boop/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file boop/requirements.txt requirements.in
#
appdirs==1.4.3            # via black
atomicwrites==1.2.1       # via pytest
attrs==18.2.0             # via black, pytest
black==18.9b0
click==7.0                # via black, pip-tools
coverage==4.5.2           # via pytest-cov
flake8==3.6.0             # via pytest-flake8
isort==4.3.4              # via pytest-isort
mccabe==0.6.1             # via flake8
mock==2.0.0
more-itertools==4.3.0     # via pytest
mypy-extensions==0.4.1    # via mypy
mypy==0.641
numpy==1.15.4
pbr==5.1.1                # via mock
pillow==5.3.0
pip-tools==3.1.0
pluggy==0.8.0             # via pytest
py==1.7.0                 # via pytest
pycodestyle==2.4.0        # via flake8
pyflakes==2.0.0           # via flake8
pytest-cov==2.6.0
pytest-flake8==1.0.2
pytest-isort==0.2.1
pytest-spec==1.1.0
pytest==4.0.0
six==1.11.0               # via mock, more-itertools, pip-tools, pytest
toml==0.10.0              # via black
typed-ast==1.1.0          # via mypy
